backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==0.5.0
psycopg2-binary==2.9.9
google-generativeai==0.8.5
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
django-allauth==65.11.0
requests==2.32.5
